pyproject.toml
[tool.poetry]
name = "fireball"
version = "0.1.0"
description = "Exploit runner"
authors = []
license = "WTFPL" # for lolz

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
aiodocker = "^0.21.0"
databases = {extras = ["sqlite"], version = "^0.4.3"}
uvicorn = {extras = ["standard"], version = "^0.14.0"}
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
mypy = "^0.910"
pylint = "^2.9.5"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
plugins = ["sqlmypy"]
